feat(BA-3349): Relocate RBAC data definitions (#7301)
