+++ changes/7301.misc.md
@@ -1,0 +1,1 @@
+Relocate RBAC type definitions for better dependency management
